requirements.txt
celery>=5.2.0
redis>=4.0.0
requests>=2.25.0
pydantic>=1.8.0
docker>=5.0.0
psutil>=5.8.0
python-dotenv>=0.19.0
